Move harness to breadboard.
